# apis/openapi.yaml
openapi: 3.0.3
info:
  title: WES — Work Planning & Release
  version: "1.0.0"
  description: >
    The core bounded context of a Warehouse Execution System (WES). It turns
    a shift's **charge** (volume due by each CPT — Critical Pull Time, the
    last moment a parcel can be manifested and still make its truck) into a
    **plan** (rate × heads per process path), releases work continuously in
    priority order (waveless — no discrete wave batches), and performs flow
    balancing off live buffer telemetry using Drum-Buffer-Rope with CPT as
    the drum. It is the "conductor": downstream of WMS planning/inventory,
    upstream of WCS equipment control.


    Aggregates owned by this service: `ChargeForecast` (volume due by CPT
    bucket), `ShiftPlan`/`PathPlan` (committed rate × heads × hours split per
    process path, enforcing `plannedHeads <= installedStations`), `WorkPool`
    (the priority queue for one process path — release-fed pools enforce a
    hard WIP limit, flow-fed pools only alarm), and `WorkUnit` (a releasable
    unit of work with at-most-once assignment and no double-complete).


    This service does **not** own equipment control (WCS), inventory
    reservation, or workforce scheduling — it only *observes* projections of
    those via Kafka integration events (`LaborPlanObserved`,
    `UsableInventoryObserved`), exposed read-only through this API. It also
    does not implement HATEOAS; this is a Richardson Maturity Level 2 API
...
  contact:
...
servers:
...
paths:
...
  /paths/{pathId}/charge:
    post:
      operationId: receiveChargeForecast
      summary: Receive a charge forecast for a process path
      description: >
        Records the volume due by CPT for a process path — the ReceiveChargeForecast
        use case. Replaces any existing forecast for the path with the new set
        of CPT buckets (each bucket is an exact CPT timestamp and the quantity
        due at that CPT). Publishes `ChargeForecastReceived`. This is the
        input side of the "charge → plan → release" pipeline: a plan is later
        committed against this charge, and work units enqueued against it are
        released against a `WorkPool` in CPT priority order.
      tags: [Charge Forecast]
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceiveChargeForecastRequest"
            example:
              buckets:
                - cpt: "2026-08-21T12:00:00Z"
                  quantity: 4200
                - cpt: "2026-08-21T18:00:00Z"
                  quantity: 1800
      responses:
        "201":
          description: >
            Charge forecast recorded. `Location` identifies the created
            forecast's canonical resource path (this service has no dedicated
            GET for a standalone charge forecast today — the total shows up
            in `GET /paths/{pathId}/telemetry` once work is enqueued and
            released against it).
          headers:
            Location:
              schema:
                type: string
              description: Canonical URI of the created charge forecast.
              example: /paths/pick-a/charge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChargeForecastResponse"
              example:
                pathId: pick-a
                totalQuantity: 6000
                buckets:
                  - cpt: "2026-08-21T12:00:00Z"
                    quantity: 4200
                  - cpt: "2026-08-21T18:00:00Z"
                    quantity: 1800
                receivedAt: "2026-08-21T08:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"
  /paths/{pathId}/plan:
    post:
      operationId: commitShiftPlan
      summary: Commit this service's shift plan for a process path
      description: >
        Commits the rate × heads × hours split for a process path — the
        CommitShiftPlan use case. Enforces the `PathPlan` aggregate invariant
        `plannedHeads <= installedStations`; violating it fails the whole
        commit (no partial commit). This is THIS service's own `ShiftPlan`
        aggregate, distinct from the labor plan Workforce Management reports
        (see `GET /paths/{pathId}/labor-plan-view`) even though both are
        called "ShiftPlan" — classic same-term-different-bounded-context.
        Publishes `ShiftPlanCommitted`.
      tags: [Shift Plan]
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitShiftPlanRequest"
            example:
              plannedHeads: 6
              installedStations: 8
              rateUnitsPerHour: 95.5
              hours: 8
      responses:
        "201":
          description: >
            Shift plan committed. `Location` identifies the committed plan's
            canonical resource path (no dedicated GET for a standalone plan
            today — `plannedThroughput` and the committed split are also
            visible indirectly via `GET /paths/{pathId}/rebalance`'s labor
            context).
          headers:
            Location:
              schema:
                type: string
              description: Canonical URI of the committed shift plan.
              example: /paths/pick-a/plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShiftPlanResponse"
              example:
                pathId: pick-a
                plannedHeads: 6
                installedStations: 8
                rateUnitsPerHour: 95.5
                hours: 8
                plannedThroughput: 4584
        "400":
          description: >
            Malformed body, a missing required field (`plannedHeads`,
            `installedStations`, `rateUnitsPerHour`, `hours` — zero-valued
            defaults are never silently accepted), an invalid field value
            (e.g. non-positive rate or hours), the `plannedHeads <=
            installedStations` invariant was violated, or the computed
            `plannedThroughput` (rate × heads × hours) is not representable
            as a finite number.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.wes-work-planning.warehouse-systems.dev/heads-exceed-installed-stations
                title: Planned heads exceed installed stations
                status: 400
                detail: "planned heads exceeds installed stations"
                instance: /paths/pick-a/plan
        "500":
          $ref: "#/components/responses/InternalError"
  /paths/{pathId}/work-units:
    post:
      operationId: enqueueWorkUnit
      summary: Enqueue a work unit onto a process path's work pool
      description: >
        Adds a releasable unit of work (e.g. a pick task) to a process path's
        `WorkPool` — the EnqueueWorkUnit use case. `workUnitId` is
        client-supplied and must be unique within the pool: enqueueing the
        same `workUnitId` twice returns `409 Conflict`
        (`work-unit-already-enqueued`) rather than silently creating a
        duplicate entry — this endpoint is idempotent-by-client-supplied-ID
        in that sense, though not a pure no-op retry (a second call with the
        same ID is rejected, not silently accepted). If this is the first
        unit ever enqueued against the path, a `WorkPool` is auto-provisioned
        as release-fed with a default WIP limit. Publishes `WorkUnitCreated`.


        **Idempotency (ADR-0022).** This resource-creating POST requires a
        caller-supplied `Idempotency-Key` header. The first request with a
        given key is processed and its response (status, headers, body — a
        4xx included) is stored in the same database transaction as the
        write; a retry with the same key and the same body replays that
        stored response verbatim and does not enqueue again. A missing key is
        `400 idempotency-key-required`; the same key with a different body
        is `422 idempotency-key-reused`. Use a fresh key per logical action.
      tags: [Work Units]
      parameters:
        - $ref: "#/components/parameters/PathId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnqueueWorkUnitRequest"
            example:
              workUnitId: wu-482910
              cpt: "2026-08-21T12:00:00Z"
              reference: order-77213-line-1
      responses:
        "201":
          description: Work unit enqueued in `Pending` state.
          headers:
            Location:
              schema:
                type: string
              description: Canonical URI of the created work unit.
              example: /work-units/wu-482910
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkUnitResponse"
              example:
                id: wu-482910
                pathId: pick-a
                cpt: "2026-08-21T12:00:00Z"
                reference: order-77213-line-1
                state: Pending
                giftWrap: false
        "400":
          description: >
            Either the `Idempotency-Key` header is missing
            (`idempotency-key-required`), or the body is malformed / fails
            validation / names an unknown `pathId` (the usual `400`
            categories).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.wes-work-planning.warehouse-systems.dev/idempotency-key-required
                title: Idempotency-Key header is required
                status: 400
                detail: "idempotency key required: this endpoint creates a new resource on every call and requires a caller-supplied Idempotency-Key header so a retried request is never applied twice"
                instance: /paths/pick-a/work-units
        "409":
          description: A work unit with this `workUnitId` is already enqueued in this pool.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.wes-work-planning.warehouse-systems.dev/work-unit-already-enqueued
                title: Work unit already enqueued in this pool
                status: 409
                detail: "work unit already enqueued in this pool"
                instance: /paths/pick-a/work-units
        "422":
          description: >
            The `Idempotency-Key` was already used with a request that had a
            different body (`idempotency-key-reused`). Use a new key for a
            genuinely different request.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.wes-work-planning.warehouse-systems.dev/idempotency-key-reused
                title: Idempotency-Key was already used with a different request
                status: 422
                detail: "idempotency key reused: the Idempotency-Key header on this request was already used with a request that had a different body; use a new Idempotency-Key for a genuinely different request"
                instance: /paths/pick-a/work-units
        "500":
          $ref: "#/components/responses/InternalError"
  /paths/{pathId}/release:
    post:
      operationId: releaseNextWork
      summary: Release the next highest-priority work unit on a process path
      description: >
        Applies the release policy to a process path's `WorkPool` — the
        ReleaseNextWork use case. Admits the pending work unit with the
        earliest (most urgent) CPT, transitioning it Pending → Released.
        This is continuous, priority-ordered admission (waveless), not a
        batch/wave release. A release-fed pool enforces its WIP limit as a
        hard invariant: releasing past the limit fails with `409`. Publishes
        `WorkReleased`, which downstream fulfillment-execution consumes to
        create a Task. This is a collection-level command on the path's work
        pool (not scoped to a single work unit id, since the caller does not
        choose which unit is released — the priority policy does), analogous
        in REST terms to a resource-scoped bulk action.
      tags: [Release]
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: The released work unit, now in `Released` state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkUnitResponse"
              example:
                id: wu-482910
                pathId: pick-a
                cpt: "2026-08-21T12:00:00Z"
                reference: order-77213-line-1
                state: Released
                giftWrap: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: No work pool has been provisioned for this path yet (no work unit has ever been enqueued).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.wes-work-planning.warehouse-systems.dev/not-found
                title: Resource not found
                status: 404
                detail: "resource not found"
                instance: /paths/pick-a/release
        "409":
          description: >
            The pool is empty (nothing pending to release) or a release-fed
            pool is already at its WIP limit.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.wes-work-planning.warehouse-systems.dev/wip-limit-reached
                title: Release-fed pool WIP limit reached
                status: 409
                detail: "WIP limit reached"
                instance: /paths/pick-a/release
        "500":
          $ref: "#/components/responses/InternalError"
  /paths/{pathId}/pool:
    put:
      operationId: configurePool
      summary: Configure a process path's work pool (feed mode and WIP limit)
      description: >
        The ConfigurePool command (ADR-0034): sets the pool's feed mode
        (`ReleaseFed` or `FlowFed`) and its WIP limit, creating the pool when
        none exists yet. This is the only way to obtain a flow-fed pool; a
        path nobody configured keeps the fallback `EnqueueWorkUnit` seeds on
        first enqueue (`ReleaseFed`, WIP limit 1000).


        **Idempotent.** `PUT` fully describes the desired configuration: the
        same body always yields the same result, and an identical repeat
        writes nothing. The alarm threshold and the pool's entries are not
        part of this command and are never touched.


        **Lowering a limit never evicts work.** If the new `wipLimit` is below
        the current `wip`, nothing is cancelled or taken back: the pool is
        simply saturated and `POST /paths/{pathId}/release` answers
        `409 wip-limit-reached` until completions bring `wip` strictly below
        the new limit. Raising the limit takes effect on the next release. The
        response reports the live `wip` so the caller can see the overshoot.
        On a `FlowFed` pool the limit is stored but not enforced (flow-fed
        pools only alarm, see ADR-0003).


        Raises no event. REST only: no MCP tool exposes this command.
      tags: [Pool Configuration]
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigurePoolRequest"
            example:
              mode: ReleaseFed
              wipLimit: 200
      responses:
        "200":
          description: The pool as configured, with its live occupancy.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolConfigResponse"
              example:
                pathId: pick-a
                mode: ReleaseFed
                wipLimit: 200
                wip: 4
                backlogDepth: 12
        "400":
          description: >
            Malformed body, a missing `wipLimit`, a `wipLimit` that is not a
            positive integer (`invalid-wip-limit`), a `mode` other than
            `ReleaseFed`/`FlowFed` (`unknown-feed-mode`), or a `pathId`
            outside the deployment's declared process-path families
            (`unknown-path-id`).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.wes-work-planning.warehouse-systems.dev/invalid-wip-limit
                title: Invalid WIP limit
                status: 400
                detail: "work pool WIP limit must be a positive integer"
                instance: /paths/pick-a/pool
        "409":
          description: >
            The pool was modified concurrently and the retry budget was
            exhausted; retry the request (`concurrent-modification`).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
              example:
                type: https://errors.wes-work-planning.warehouse-systems.dev/concurrent-modification
                title: The work pool was modified concurrently; retry the request
                status: 409
                detail: "concurrent modification"
                instance: /paths/pick-a/pool
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  parameters:
    PathId:
      name: pathId
      in: path
      required: true
      description: >
        The process path's identifier — a named station that owns a queue
        (unit-in → unit-out), e.g. a pick zone or pack line. Must belong to
        a process-path family declared in this deployment's process-path
        catalogue (ADR-0012; today the fleet's sortable-fc topology
        declares the pick/pack/rebin/slam families, each optionally
        carrying a "-"-delimited station/zone suffix, e.g. pick-a or
        pack-station-1). The write endpoints (receiveChargeForecast,
        commitShiftPlan, enqueueWorkUnit) reject ids outside the declared
        families with `400 unknown-path-id`; read endpoints surface a
        missing pool/projection as `404` instead.
      schema:
        type: string
        minLength: 1
        pattern: '^(pick|pack|rebin|slam)(-.*)?$'
      example: pick-a
...
  schemas:
...
    ReceiveChargeForecastRequest:
      type: object
      properties:
        buckets:
          type: array
          minItems: 1
          description: At least one CPT/quantity bucket is required.
          items:
            $ref: "#/components/schemas/CPTBucket"
      required: [buckets]
...